+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+TBD
+* Adding Get Device Token API for shared device mode #2980
+* Improve UI tests performance #2981
+
 ## [2.11.0]:
 * Rename file on disk to match Xcode #2909
 * Provide public api to set custom headers for CIAM requests #2862